Made changes to testing and to querying inside of get_courses()
Tests now have exceptions internal to the project that follow the python exception naming convention when instead of ending the name with "exception" it ends with "error". Also made the pitt server error extend from time out error.

Added license to all test files.

Cleaned names of test functions since part of the name is redundant since the class already tells which module it's coming from.

Added a constant list of programs internal to the Pitt API, it needs some reworking and checking but current state is pretty good. Also get_courses() now raises a ValueError when an invalid subject is entered and also now returns an empty list when no courses are offered for a certain subject.