New functionality with extract_id and updated _find_item
Extract ID is now able to use either instructor or section, or both if one fails and the other doesn't. It also will give the proper error message based on what arguments are not none.
Find item has a new error_item parameter for elaborating which item it can't find in the error message.