Added Library API and its tests
Allows document search by general query or bookmark. A general query will
include data about the distribution of search results with regard to fields,
conent type, language, and so on. It will also include the number of search
results, pages, and the current page, to allow for search by page.