Cleaned up how Textbook tests fetch json
Switch from continuously fetching data from file and instead in the setup of the class just fetch once and store internally.
In the future the json samples will be slimmed down since majoriy of the data isn't need in the file. But since we have a sample response we may be able to do more in depth testing making sure it's capable of fetching and digesting entire data of textbook data.