Fix mypy errors in laundry.py
Mypy raised errors when a variable is used to hold two different types. These
were fixed by introduction a new variable instead of reassigning the old one.

The type annotation for get_status_detailed is changed to reflect that
'time_left' is cast to an integer. In get_status_simple it is still returned
as a string, which is inconsistent, but I will leave this decision up to
someone who knows the API better.