Add type annotations to library.py
Make some minimal changes to fix mypy errors. One common type is when a
variable is reassigned.  In another case, mypy wanted a variable to be
explicitly annotated.

I also fixed an error (not sure if an actual funcitonal error, but it was
flagged by mypy) where the & operator was used with two incompatible types.