Refactoring course api
Moved off of beautifulsoup which allowed a gain in more readable and maintainable code.
Added requests_html. From this usage, i'm planning to phase out beautiful in favor of this library.
Add parse too. It comes with requests_html so not really an extra dep. but has some nice features for reducing usage of some regex